LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/controlled-dedicatedworker-postMessage.https-expected.txt


PASS Register service worker
FAIL Verify dedicated worker gets messages if setting event listener early assert_equals: expected (number) 0 but got (string) "No message received"
PASS Verify dedicated worker does not get all messages if not setting event listener early
